Renames xmpp.js to Connection.js to reflect the embedded object name
